gym_bot/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pycryptodomex>=3.20.0
pyyaml>=6.0
playwright>=1.40.0    # 仅滑块验证码降级时需要
